Cache StoragePods
They're loaded for every datastore used in the cluster.
